// api/generated/model/exerciseFormDtoBodyPart.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type ExerciseFormDtoBodyPart = typeof ExerciseFormDtoBodyPart[keyof typeof ExerciseFormDtoBodyPart];


export const ExerciseFormDtoBodyPart = {
  Unknown: 'Unknown',
  Chest: 'Chest',
  Back: 'Back',
  Shoulders: 'Shoulders',
  Biceps: 'Biceps',
  Triceps: 'Triceps',
  Forearms: 'Forearms',
  Abs: 'Abs',
  Quads: 'Quads',
  Hamstrings: 'Hamstrings',
  Calves: 'Calves',
  Glutes: 'Glutes',
} as const;